requirements.txt
asgiref==3.11.0
Django==5.2.9
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.4
typing_extensions==4.15.0
drf-spectacular
django-cors-headers

